+++ packages/waku/src/lib/builder/serve-netlify.ts
@@ -1,7 +1,7 @@
 import { Hono } from 'hono';
 import type { Context } from '@netlify/functions';
 
-import { honoMiddleware } from '../middleware/hono-prd.js';
+import { honoMiddleware } from '../old-wrappers/hono-prd.js';
 
 const ssr = !!import.meta.env.WAKU_BUILD_SSR;
 const loadEntries = () => import(import.meta.env.WAKU_ENTRIES_FILE!);
