+++ packages/waku/src/lib/builder/serve-partykit.ts
@@ -1,6 +1,6 @@
 import { Hono } from 'hono';
 
-import { honoMiddleware } from '../middleware/hono-prd.js';
+import { honoMiddleware } from '../old-wrappers/hono-prd.js';
 
 const ssr = !!import.meta.env.WAKU_BUILD_SSR;
 const loadEntries = () => import(import.meta.env.WAKU_ENTRIES_FILE!);
