+++ packages/waku/src/prd.ts
@@ -1,4 +1,2 @@
-export { honoMiddleware as unstable_honoMiddleware } from './lib/middleware/hono-prd.js';
-export { connectMiddleware as unstable_connectMiddleware } from './lib/middleware/connect-prd.js';
-
-export { createHandler as unstable_createHandler } from './lib/handlers/handler-prd.js';
+export { honoMiddleware as unstable_honoMiddleware } from './lib/old-wrappers/hono-prd.js';
+export { connectMiddleware as unstable_connectMiddleware } from './lib/old-wrappers/connect-prd.js';
